added oracles
