+++ contracts/liquidity_book/lb_pair/src/lib.rs
@@ -4,4 +4,7 @@ mod error;
 mod prelude;
 
 pub mod contract;
+pub mod execute;
+pub mod helper;
+pub mod query;
 pub mod state;
