+++ packages/shade_protocol/src/contract_interfaces/liquidity_book/lb_lib_todo_list.md
@@ -1,0 +1,201 @@
+- bin_helper.rs
+  - assert_approxeq_abs
+  - decode
+  - get_amount_out_of_bin
+  - get_amounts
+  - get_composition_fees
+  - get_liquidity
+  - get_shares_and_effective_amounts_in
+  - is_empty
+  - received
+  - received_amount
+  - received_x
+  - received_y
+  - transfer
+  - transfer_x
+  - transfer_y
+  - verify_amounts
+- constants.rs
+- error.rs
+- fee_helper.rs
+  - get_composition_fee
+  - get_fee_amount
+  - get_fee_amount_from
+  - get_protocol_fee_amount
+  - verify_fee
+  - verify_protocol_share
+- mod.rs
+  - approx_div
+- oracle_helper.rs
+  - bound
+  - update
+- pair_parameter_helper.rs
+  - get_active_id
+  - get_base_factor
+  - get_base_fee
+  - get_decay_period
+  - get_delta_id
+  - get_filter_period
+  - get_id_reference
+  - get_max_volatility_accumulator
+  - get_oracle_id
+  - get_protocol_share
+  - get_reduction_factor
+  - get_time_of_last_update
+  - get_total_fee
+  - get_variable_fee
+  - get_variable_fee_control
+  - get_volatility_accumulator
+  - get_volatility_reference
+  - set_active_id
+  - set_oracle_id
+  - set_static_fee_parameters
+  - set_volatility_accumulator
+  - set_volatility_reference
+  - update_id_reference
+  - update_references
+  - update_time_of_last_update
+  - update_volatility_accumulator
+  - update_volatility_parameters
+  - update_volatility_reference
+- price_helper.rs
+  - convert128x128_price_to_decimal
+  - convert_decimal_price_to128x128
+  - get_base
+  - get_exponent
+  - get_id_from_price
+  - get_price_from_id
+- transfer.rs
+  - space_pad
+  - to_cosmos_msg
+- types.rs
+- viewing_keys.rs
+  - as_bytes
+  - check_viewing_key
+  - create_hashed_password
+  - create_viewing_key
+  - fmt
+  - from
+  - new
+  - register_receive
+  - set_viewing_key_msg
+  - to_hashed
+- lb_token
+  - expiration.rs
+    - default
+    - fmt
+    - is_expired
+  - metadata.rs
+  - mod.rs
+  - permissions.rs
+    - check_view_balance_perm
+    - check_view_pr_metadata_perm
+  - state_structs.rs
+    - default_fungible
+    - default_nft
+    - flatten
+    - to_enum
+    - to_store
+  - txhistory.rs
+    - into_humanized
+- math
+  - bit_math.rs
+    - closest_bit_left
+    - closest_bit_right
+    - least_significant_bit
+    - most_significant_bit
+  - encoded_sample.rs
+    - decode
+    - decode_bool
+    - decode_uint12
+    - decode_uint128
+    - decode_uint14
+    - decode_uint16
+    - decode_uint20
+    - decode_uint24
+    - decode_uint40
+    - decode_uint64
+    - decode_uint8
+    - set
+    - set_bool
+  - liquidity_configurations.rs
+    - get_amounts_and_id
+    - new
+    - update_distribution
+  - mod.rs
+  - packed_u128_math.rs
+    - add
+    - add_alt
+    - decode
+    - decode_alt
+    - decode_x
+    - decode_y
+    - encode
+    - encode_alt
+    - encode_first
+    - encode_second
+    - gt
+    - lt
+    - max
+    - min
+    - scalar_mul_div_basis_point_round_down
+    - sub
+    - sub_alt
+  - sample_math.rs
+    - encode
+    - get_cumulative_bin_crossed
+    - get_cumulative_id
+    - get_cumulative_txns
+    - get_cumulative_volatility
+    - get_fee_token_x
+    - get_fee_token_y
+    - get_sample_creation
+    - get_sample_last_update
+    - get_sample_lifetime
+    - get_vol_token_x
+    - get_vol_token_y
+    - get_weighted_average
+    - set_created_at
+    - update
+  - tree_math.rs
+    - add
+    - _closest_bit_left
+    - _closest_bit_right
+    - contains
+    - default
+    - find_first_left
+    - find_first_right
+    - new
+    - remove
+  - u128x128_math.rs
+    - log2
+    - pow
+  - u24.rs
+    - add
+    - cmp
+    - div
+    - eq
+    - fmt
+    - mul
+    - new
+    - partial_cmp
+    - sub
+    - value
+  - u256x256_math.rs
+    - addmod
+    - _get_end_of_div_round_down
+    - _get_mul_prods
+    - mul_div_round_down
+    - mul_div_round_up
+    - mulmod
+    - mul_shift_round_down
+    - mul_shift_round_up
+    - shift_div_round_down
+    - shift_div_round_up
+    - u256_to_u512
+    - u512_to_u256
+  - uint256_to_u256.rs
+    - split_u256
+    - split_uint256
+    - u256_to_uint256
+    - uint256_to_u256
